requirements.txt
python-telegram-bot==20.7
mysql-connector-python
python-dotenv

